Build fixes

+++ src/native/libs/System.Native/pal_io.c
@@ -352,7 +352,7 @@ int32_t SystemNative_Close(intptr_t fd)
 
 intptr_t SystemNative_Dup(intptr_t oldfd)
 {
-    int result;
+    intptr_t result;
 #if HAVE_F_DUPFD_CLOEXEC
     while ((result = fcntl(ToFileDescriptor(oldfd), F_DUPFD_CLOEXEC, 0)) < 0 && errno == EINTR);
 #elif HAVE_F_DUPFD
